requirements.txt
appdirs==1.4.3
astroid==2.3.3
attrs==19.3.0
black==19.10b0
click==7.1.1
Cython==0.29.16
Flask==1.1.2
Flask-WTF==0.14.3
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
pathspec==0.8.0
Pillow==7.1.1
pkg-resources==0.0.0
pylint==2.4.4
pytesseract==0.3.4
regex==2020.4.4
six==1.14.0
tesseract==0.1.3
tesseract-ocr==0.0.1
toml==0.10.0
typed-ast==1.4.1
Werkzeug==1.0.1
wrapt==1.11.2
WTForms==2.2.1
